Checks all sources in for troubleshooting MS Validator error in Italic styles:
Message: E1703
--------------------------------------------------------------------------------
loca: Loca entry points outside the glyf range

There is an entry in the loca table that points outside the range(s) of glyf data. This may happen if the loca table has not been updated to reflect the removal of the final glyph(s) from a font.